Use Weirich Brace Convention for block delimiters
The `do/end` vs. `{}` rule for single- vs. multi-line blocks requires rewriting
whenever the code in the block changes for unrelated reasons.

Instead, a more consistent rule exists to indicate the purpose of the block:

- Use `{}` for FUNCTIONAL blocks -- when the return value of the block is used
- Use `do/end` for PROCEDURAL blocks -- when the return value is discarded

This makes chaining of blocks more natural with the `{}` syntax, whether they
span over multiple lines or not.

It also works out that (a lot of the time) one-liners are functional while
procedures span over lines, so it looks almost the same as the line-number rule.

Specific methods are predefined as ProceduralMethods (`each`, `before`,
`configure`, etc.) or FunctionalMethods (`map`, `select`, `let`, etc.) and can
be configured to add more if needed.