Don't leave trailing commas in method arguments
Method calls are fundamentally different from collection values like hashes &
arrays. We can normally expect that collections will have items added to them so
the comma rule makes sense there, but the same is rarely true of method
arguments: argument lists should never have so many items that this is needed.
It's actually more helpful to leave off the comma indicating the last argument
in this case.

This rule also complains when the argument is a single item, chained over
multiple lines. The comma does not add clarity here:

    where(
      something_long
        .map { |a| a.even_longer }
        .reduce(:or),
    )