Reduce the minimum allowed param length from 3 to 2
When the author provides more than one character for a param name, it indicates
that they have at least put some thought into it. This allows more 2 letter
names such as "op", "db", etc.

The override for block param names is removed because the default there is 1
character and it's redundant.